+++ src/module/SummonMenu/app.svelte
@@ -81,7 +81,14 @@
 				throw ui.notifications.warn("There is no active GM to accept this summon!");
 			}
 			foundryApp.minimize();
-			await pick({ uuid, updateData: { prototypeToken: { actorLink: data.options.actorLink }, ...data.options.updateData }});
+
+			const token = await pick({
+				uuid,
+				updateData: { prototypeToken: { actorLink: data.options.actorLink },
+				...data.options.updateData
+			}});
+
+			if (token) foundryApp._completeSelection(token);
 			if (data.options.once) foundryApp.close();
 		} catch {
 			foundryApp.maximize();